refactor: eject from next-mdx-remote
