+++ app/posts/[slug]/actions.ts
@@ -4,8 +4,6 @@ import path from 'path'
 
 import { notFound } from 'next/navigation'
 
-import { fetchPublishedPosts } from '../actions'
-
 import { getPostFromMDX } from '@/lib/mdx'
 import type { Post } from '@/lib/types'
 
@@ -21,15 +19,4 @@ export async function fetchPostBySlug(slug: string): Promise<Post> {
       throw new Error(`Something went wrong. Unable to fetch a blog post for ${slug}`)
     }
   }
-}
-
-export async function fetchPreviousPostSlug(slug: string): Promise<string | undefined> {
-  const publishedPosts = await fetchPublishedPosts()
-  const postIndex = publishedPosts.findIndex((post) => post.slug === slug)
-
-  if (postIndex === publishedPosts.length - 1) return undefined
-
-  const previousPost = publishedPosts.at(postIndex + 1)
-
-  return previousPost && previousPost.slug
 }
