--- next.config.mjs
@@ -1,24 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-  images: {
-    remotePatterns: [],
-  },
-  transpilePackages: ['next-mdx-remote'],
-  async rewrites() {
-    return [
-      {
-        source: '/ingest/static/:path*',
-        destination: 'https://us-assets.i.posthog.com/static/:path*',
-      },
-      {
-        source: '/ingest/:path*',
-        destination: 'https://us.i.posthog.com/:path*',
-      },
-    ]
-  },
-  // ? NOTE: This is required to support PostHog trailing slash API requests
-  skipTrailingSlashRedirect: true,
-}
-
-export default nextConfig
