+++ package.json
@@ -13,7 +13,10 @@
     }
   },
   "lint-staged": {
-    "*.{ts,tsx}": ["next lint --fix", "pnpm format"],
+    "*.{ts,tsx}": [
+      "next lint --fix",
+      "pnpm format"
+    ],
     "*.md": "prettier --list-different"
   },
   "scripts": {
@@ -39,6 +42,9 @@
     "typecheck": "tsc --pretty"
   },
   "dependencies": {
+    "@mdx-js/loader": "3.1.0",
+    "@mdx-js/react": "3.1.0",
+    "@next/mdx": "15.3.5",
     "@radix-ui/react-accordion": "1.2.11",
     "@radix-ui/react-aspect-ratio": "1.1.7",
     "@radix-ui/react-avatar": "1.1.10",
@@ -56,7 +62,6 @@
     "geist": "1.4.2",
     "lucide-react": "0.511.0",
     "next": "15.3.3",
-    "next-mdx-remote": "5.0.0",
     "next-themes": "0.4.6",
     "posthog-js": "1.255.0",
     "posthog-node": "4.17.2",
@@ -64,7 +69,6 @@
     "react-dom": "19.1.0",
     "rss": "1.2.2",
     "sharp": "0.34.2",
-    "shiki": "3.4.2",
     "tailwind-merge": "3.3.1"
   },
   "devDependencies": {
@@ -74,9 +78,11 @@
     "@next/eslint-plugin-next": "15.3.4",
     "@playwright/test": "1.53.0",
     "@tailwindcss/postcss": "4.1.11",
+    "@tailwindcss/typography": "0.5.16",
     "@testing-library/jest-dom": "6.6.3",
     "@testing-library/react": "16.3.0",
     "@testing-library/user-event": "14.6.1",
+    "@types/mdx": "2.0.13",
     "@types/node": "22.7.5",
     "@types/react": "19.1.8",
     "@types/react-dom": "19.1.5",
@@ -104,7 +110,10 @@
     "rehype-autolink-headings": "7.1.0",
     "rehype-pretty-code": "0.14.1",
     "rehype-slug": "6.0.0",
+    "remark-frontmatter": "5.0.0",
     "remark-gfm": "4.0.1",
+    "remark-mdx-frontmatter": "5.2.0",
+    "shiki": "3.4.2",
     "tailwindcss": "4.1.10",
     "tailwindcss-animate": "1.0.7",
     "tw-animate-css": "1.3.0",
